ollama-docker-app/SelfDB/tests/requirements.txt
requests
WebSockets
python-dotenv